vanna/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
psycopg[binary]==3.2.3
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6

# Alternative: asyncpg (for Linux/production environments)
# asyncpg==0.30.0
